[OSDEV-986] Fix the population of the custom data points uploaded via lists. (#157)
Correct the population of the custom data points that are uploaded via
lists. Ensure that the full list header is saved in the database, and
that the raw data for each facility list item is saved as a string of
strings, with each value separated by a comma. This way, it helps
maintain backward compatibility with the functionality responsible for
displaying custom data points on the embedded maps. Also, revert to the
previous default logic, which saves the sector as `Unspecified` when
sector, sector_product_type, or product_type have empty values.