Decrease the number of containers for prod, preprod and test environments (#179)
- 1.12 changes: Resource allocation has been optimized for the Test
environment. The number of ECS tasks in the Test environment has been
reduced from 4 to 2, while maintaining system stability.

- Back merge of the 1.11 changes: The production environment seems
stable enough even with fewer workers/containers than usual. We've been
running with 12 for the past two days, so there's no need to spend money
on 16. That's why we've decided to decrease the number of containers for
the prod and preprod environments to 12.