[OSDEV-931] Facility Claims. Use account email for contact info on claim request.   (#182)
The following features have been implemented:
- Made the Email field in the claim form uneditable, setting the
claimer's email as the default value for this field.
- Removed the _Preferred method of contact_ field from both the claim
form and the claim details page in the admin dashboard.
- Since email and preferred_contact_method are no longer necessary
for the claim form, they have been removed from the FacilityClaim
model and the respective history table. Additionally, the old
index_approved_claim function has been replaced with a similar one
that does not index the preferred_contact_method field.