[OSDEV-997] Facility Claims. Send messages to claimant from the claim request page. (#189)
[OSDEV-997](https://opensupplyhub.atlassian.net/browse/OSDEV-997) -
Facility Claims. Send messages to claimant from the claim request page.

A new button, 'Message Claimant' has been added to the update status
controls on the Facility Claim Details page. After successfully sending
a message, the message text is recorded in the Claim Review Notes.
A new method, `message_claimant`, was added to the
`FacilityClaimViewSet` for handling a POST request on the url-path
`message-claimant` for messages to the claimant.
Mail templates for the message to the claimant and the claims team
signature were also added.