[OSDEV-878] Create post-deployment django command (#222)
[OSDEV-878](https://opensupplyhub.atlassian.net/browse/OSDEV-878) Create
post-deployment django command

Added a Django command `post_deployment` that runs Django migrations
during the deployment process. This command can be expanded to include
other post-deployment tasks.
Used the `post_deployment` command in the `post_deploy` job of the
Deploy to AWS workflow.

The implementation of this command allows for the automation of
post-deployment scenarios when more than just migrations are necessary.