[OSDEV-1019] Login page. User did not click confirmation link. (#228)
[OSDEV-1019](https://opensupplyhub.atlassian.net/browse/OSDEV-1019)
Login page. User did not click confirmation link.

Fixed an error message to 'Your account is not verified. Check your
email for a confirmation link.' when a user tries to log in with an
uppercase letter in the email address and their account has not been
activated through the confirmation link.