[OSDEV-1084] Claims. Update logic for number of workers. (#231)
[OSDEV-1084](https://opensupplyhub.atlassian.net/browse/OSDEV-1084)
Claims. Update logic for number of workers.

The implementation of this issue makes it possible to add a range for
the number of workers during the claiming process, either after pressing
the “I want to claim this production location” link or on the Claimed
Facility Details page.

To change the type of the `facility_workers_count` field in the
`FacilityClaim` table from `IntegerField` to `CharField`, there were
added migrations with the next sequence:
* 0146_add_facility_workers_count_new_field_to_facilityclaim - adds the
facility_workers_count_new field to the FacilityClaim model.
* 0147_copy_facility_workers_count_to_facility_workers_count_new -
copies the data from the facility_workers_count field to the
facility_workers_count_new field.
* 0148_remove_facility_workers_count_field_from_facilityclaim - removes
the facility_workers_count field from the FacilityClaim model.
* 0149_rename_facility_workers_count_new_to_facility_workers_count -
renames the facility_workers_count_new field to facility_workers_count.