[OSDEV-705] Upload Error. "Could not convert string to float" for empty coordinate cells. (#272)
[OSDEV-705](https://opensupplyhub.atlassian.net/browse/OSDEV-705) Upload
Error. "Could not convert string to float" for empty coordinate cells.

Fixed the error “could not convert string to float” that occurred when a
list contained columns for “lat” and “lng” and only some of the rows in
these columns had data. As a result, rows are processed regardless of
whether the values for “lat” and “lng” are present and valid, invalid,
or empty.
Created an additional `RowCoordinatesSerializer` in the ContriCleaner to
handle coordinate values ("lat" and "lng"). Moved the conversion of
"lat" and "lng" into float point numbers from `FacilityListViewSet` to
this serializer.


[OSDEV-705]:
https://opensupplyhub.atlassian.net/browse/OSDEV-705?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ