Disable using legacy template endpoint for OpenSearch (#296)
The legacy _template API endpoint was disabled via the configuration file in favor of the new _index_template API endpoint, since the composable index template is used for OpenSearch. The legacy_template was set to false to start using the defined composable index template in the production_locations.json file. See the documentation about the legacy_template configuration parameter. This change is necessary to avoid omitting the production_locations.json index template for the production-locations index defined in the Logstash app and to enforce the OpenSearch cluster to use the explicit mapping for the production-locations index.