[OSDEV-1005] Propagate facility deletions to OpenSearch data source (#267)
Propagate facility deletions to OpenSearch data source

- Implement the propagation of production location deletions from the
PostgreSQL database to the OpenSearch cluster. After this fix, the
locations that were deleted will be excluded from the response of the
`v1/production-location` GET API endpoint.
- Enable deployment of the Logstash and OpenSearch infra to AWS
environments.
- Disconnect location deletion propagation to the OpenSearch cluster
while the Django tests are running, as it is outside the scope of Django
unit testing.