[OSDEV-1039] make list parsing asynchronous (#356)
[OSDEV-1039] - Made the list parsing asynchronous and increased the list upload limit
to 10,000 facilities per list to reduce manual work for moderators when
they split large lists into smaller ones. The following architectural
and code changes have been made:

1. Renamed the previously copied `api/facility-lists/createlist` POST
endpoint to the `api/facility-lists` POST endpoint. Deleted the old
implementation of the `api/facility-lists` POST endpoint along with the
`use_old_upload_list_endpoint` switcher that was necessary to toggle
between the old and new list upload endpoints.
2. Removed the triggering of ContriCleaner from the `api/facility-lists`
POST endpoint and moved it to the async parse AWS batch job to reduce
the load on the endpoint. Introduced a `parsing_errors` field for the
FacilityList model to collect list-level and internal errors logged
during the background parsing of the list.
3. Established a connection between the EC2 instance within the AWS
batch job and the S3 bucket where all the uploaded list files are saved.
This is necessary because the parse job retrieves a particular list from
the S3 bucket via Django.
4. Deleted redundant code from the previous implementation of the list
item parsing.
5. Adjusted Django, ContriCleaner, and integration tests. Regarding
integration tests, the `facility_list_items.json` fixture was converted
to concrete CSV lists, which were connected to the `facility_lists.json`
fixture to upload them to the DB while creating the test DB for the
integration tests. This is necessary because the parsing function that
triggers ContriCleaner can only work with real files, not facility list
items as it did previously.
6. Refactored the ContributeForm component in the front-end app.
7. The list page was adjusted to work with asynchronous parsing, and a
new dialog window was added to notify users about the list parsing
process, indicating that they need to wait.
8. Introduced a UI to display list parsing errors on the list page after
the page refresh.