Move all the files related to the OpenSearch service to the existing OpenSearch folder (#406)
Moved all the files related to the OpenSearch service to the existing
`src/django/api/services/opensearch` folder within the `api` app of the
Django application. This should make it easier to navigate through the
files and clarify the location of all OpenSearch service-related files
in one place within the `api` app in Django.