[OSDEV-1170] Create database dump from the latest anonymized snapshot. (#373)
This pull request aims to simplify and automate the process of
transferring data from the Production environment to the Test and
Pre-Prod environments. The new functionality enables the automatic
generation of a dump with up-to-date data from the latest Production
database snapshot, providing a convenient and efficient way to refresh
test environments with real (anonymized) data. This ensures more
relevant data for testing new features and changes.