Right-sized ECS and RDS resources (#458)
Right-sized the resources for Django containers across all environments
and the RDS instance in the Production and Preprod environments. This
will result in a savings of approximately $2,481. The following changes
have been made:
 - Production:
 - RDS instance type was changed from `db.m6in.8xlarge` to
`db.m6in.4xlarge`.
  - ECS tasks for Django containers: the number was reduced from `12` to
`10`, and memory was reduced from `8GB` to `4GB`.
 - Preprod:
  - RDS instance type was changed from `db.m6in.8xlarge` to
`db.m6in.4xlarge`.
  - ECS tasks for Django containers: the number was reduced from `12` to
`10`, and memory was reduced from `8GB` to `4GB`.
  - These changes were made to maintain parity with the Production
environment, as it is a copy of that environment.
 - Staging:
  - ECS tasks for Django containers: memory was reduced from `8GB` to
`2GB`.
 - Test:
  - ECS tasks for Django containers: memory was reduced from `8GB` to
`4GB`.
 - Development:
  - ECS tasks for Django containers: memory was reduced from `8GB` to
`1GB`, and CPU was reduced from `1 vCPU` to `0.5 vCPU`.