[OSDEV-1336] SLC. Implement PATCH api/v1/production-locations/{os_id}/ (#457)
- Introduced a new PATCH `/api/v1/production-locations/{os_id}/`
endpoint based on the API v1 specification. This endpoint allows the
creation of a new moderation event for updating the production location
with the given details. Basically, the endpoint can be used to
contribute to an existing location.
- Dynamic mapping for the new fields in the `moderation-events` index
has been disabled for those that don't have an explicit mapping defined.
This change helps avoid indexing conflicts, such as when a field is
initially indexed with one data type (e.g., long), but later an entry
with a different data type for the same field is indexed, causing the
entire entry to fail indexing. After this change, fields with an
explicit mapping will be indexed, while other fields will not be indexed
or searchable, but will still be displayed in the document. The
`clear_opensearch.sh.tpl` script has been edited to clear only resources
related to the `moderation-events` index, so there will be no need to
wait for the production-locations index to be refilled during
deployment, as it will not be deleted.