[OSDEV-1409] Implement PATCH /api/moderation-events/{moderation_id}/production-locations/{os_id}/ endpoint (#438)
[OSDEV-1409](https://opensupplyhub.atlassian.net/browse/OSDEV-1409) -
Moderation. Implement PATCH
api/moderation-events/{modreation_id}/production-locations/{os_id}/
endpoint.

In this PR was introduced a new `PATCH
/api/moderation-events/{moderation_id}/production-locations/{os_id}/`
endpoint. This endpoint allows the creation of a new contribution for an
existing production location based on the provided moderation event.

While executing this endpoint, the `api_facilitylistitem` record is
created with the status **"CONFIRMED_MATCH"**.
The `api_facilitymatch` record is created with the status
**"CONFIRMED"**, a confidence value of **1**, and the following result:
`{"match_type": "moderation_event_confirmed_match"}`.

This behavior simulates a scenario in which DedupeHub returns a single
potential match that is subsequently confirmed manually by the
moderator.

[OSDEV-1409]:
https://opensupplyhub.atlassian.net/browse/OSDEV-1409?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: vladsha-dev <newshapovalov2002@gmail.com>
Co-authored-by: Vadim Kovalenko <vadim.kovalenko@speedandfunction.com>
Co-authored-by: Roman Stolar <roman.stolar@speedandfunction.com>
Co-authored-by: Vadim Kovalenko <90321493+VadimKovalenkoSNF@users.noreply.github.com>