[OSDEV-1581] Add Geohex grid aggregation support to the production locations endpoint (#488)
[OSDEV-1581](https://opensupplyhub.atlassian.net/browse/OSDEV-1581) -
[API] Add Geohex grid aggregation support to the production locations
endpoint.

Added Geohex grid aggregation support to the GET
`/api/v1/production-locations/` endpoint. To receive the Geohex grid
aggregation list in the response, it is necessary to pass the
`aggregation` parameter with a value of `geohex_grid` and optionally
specify `geohex_grid_precision` with an integer between 0 and 15. If
`geohex_grid_precision` is not defined, the default value of 5 will be
used.

Added test cases to cover the Geohex grid aggregation support.

[OSDEV-1581]:
https://opensupplyhub.atlassian.net/browse/OSDEV-1581?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ