Updated the prefix from "stg" to "dev" for the Development environment in the AWS infrastructure. (#493)
Some of the resources related to the Development AWS environment still
have the `stg` prefix, which can be confusing since we also have a
Staging environment with the same prefix. To clarify the resource names,
including the database instance, the prefix has been updated from `stg`
to `dev` for the development environment.