[OSDEV-1515] Database. Upgrade the PostgreSQL version to 16. (#490)
- Upgraded the PostgreSQL version from 13 to 16 for the database used in
local development, DB anonymization, DB restore setup, and environments
in the AWS cloud. Additionally, the pg_trgm extension has been upgraded
to version 1.6 based on the available extension version for PostgreSQL
16.3 in AWS RDS.
- 0163_refresh_pg_statistic_and_upgrade_postgres_extensions.py - Updated
the SQL script within the migration that upgrades the DB extension
versions to handle previously failure cases when a higher version is
available for upgrade or when the extension is not installed. This is
primarily useful for local development or DB resets in the Development
environment, where migrations are applied from scratch, one by one. This
fix will not negatively affect other environments, as the migration has
already been applied and will not be reapplied. Additionally, the
changes are backward compatible.
- 0164_refresh_pg_statistic_and_upgrade_postgres_extensions_after_db_upgrade_to_postgres_16.py
- Created the migration that refreshes the `pg_statistic` table after
the upgrade to PostgreSQL 16 and upgrades the pg_trgm extension to
version 1.6. The SQL script within the migration that upgrades the DB
extension versions handles previously failure cases where a higher
version is available for upgrade or where the extension is not
installed.