[OSDEV-899] Move React application to the S3 bucket (#449)
With this task, we split the Django container into two components: FE
(React) and BE (Django). Requests to the frontend (React) will be
processed by the CDN (CloudFront), while requests to the API will be
redirected to the Django container. This approach will allow for more
efficient use of ECS cluster computing resources and improve frontend
performance.

  The following endpoints will be redirected to the Django container:
  * tile/*
  * api/*
  * /api-auth/*
  * /api-token-auth/*
  * /api-feature-flags/*
  * /web/environment.js
  * /admin/*
  * /health-check/*
  * /rest-auth/*
  * /user-login/*
  * /user-logout/*
  * /user-signup/*
  * /user-profile/*
  * /user-api-info/*
  * /admin
  * /static/admin/*
  * /static/django_extensions/*
  * /static/drf-yasg/*
  * /static/gis/*
  * /static/rest_framework/*
  * /static/static/*
  * /static/staticfiles.json

  All other traffic will be redirected to the React application.

---------

Co-authored-by: vladsha-dev <newshapovalov2002@gmail.com>