[OSDEV-1782] Create a pop-up window to enter a text message for the rejection action of contribution. (#538)
[OSDEV-1782](https://opensupplyhub.atlassian.net/browse/OSDEV-1782) -
[SLC] Moderation Queue: Record rejection options & enable pop-up window
to enter text.

In this PR, the following changes were made:

- New fields action_reason_text_cleaned and action_reason_text_raw were
added to the api_moderationevent table to store text messages received
when a moderator takes an action on a moderation event.
- Additional validation was added for the fields
action_reason_text_cleaned and action_reason_text_raw when using the
PATCH api/v1/moderation-events/{moderation_id} endpoint. These fields
are required in the request body when the status field is set to
'REJECTED'. The minimum length for the values of these fields is 30
characters.
- A confirmation dialog window was added that appears when a user tries
to reject a moderation event. The dialog includes a WYSIWYG text editor
where entering a message of at least 30 characters is required to
confirm the rejection. If a user does not enter the required number of
characters, the 'Reject' button is disabled, and a tooltip with a clear
message appears when the mouse hovers over it."
- Was added sanitization on the server side by using the `Django-Bleach`
library for the HTML content that is stored in the
`action_reason_text_raw` field. The bleach filter was applied to the
`action_reason_text_raw` value in the
`slc_contribution_rejected_body.html` template.

![Screenshot from 2025-03-04
22-21-09](https://github.com/user-attachments/assets/0118605d-b99c-43e3-845a-e29b3af12c19)
![Screenshot from 2025-03-04
19-51-02](https://github.com/user-attachments/assets/fc9f8eb3-5393-4c36-9693-d59369d6e6ac)



[OSDEV-1782]:
https://opensupplyhub.atlassian.net/browse/OSDEV-1782?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ