[OSDEV-1838] Fix search by OS ID that contains slash or other special characters. (#554)
[OSDEV-1838](https://opensupplyhub.atlassian.net/browse/OSDEV-1838) -
SLC. Fix search by OS ID that contains slash.

This PR fixes an issue where the router redirected to an unsupported
page when the OS ID contained a forward slash. The issue was resolved by
encoding the OS ID value using the `encodeURIComponent()` function
before passing it as a URL parameter.

![Screenshot from 2025-03-10
22-14-33](https://github.com/user-attachments/assets/65eb8adc-1db1-4d1f-9232-b9c0d0227680)
![Screenshot from 2025-03-10
22-15-14](https://github.com/user-attachments/assets/b2ce1a8f-8c88-412b-b001-ef186fc3c5f3)


[OSDEV-1838]:
https://opensupplyhub.atlassian.net/browse/OSDEV-1838?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ