[OSDEV-1930] Implemented displaying additional identifiers such as RBA, LEI, and DUNS IDs on the production location profile page. (#594)
- Implemented front-end logic to display additional identifiers such as
RBA, LEI, and DUNS IDs as data points on the production location profile
page, once the `show_additional_identifiers` feature flag is returned
with a true value from the backend.
- 0168_introduce_show_additional_identifiers_switch.py - This migration
introduces a new switch called `show_additional_identifiers`, which will
be used on the production location profile page to show or hide
additional identifiers of the production location.
- Updated the styles of primary and secondary text in the data points UI
and the location details contributor drawer on the location profile
page, as well as the hash in the React error boundary component, to
prevent overflow on the page or within its field location. Replaced the
deprecated `word-wrap` CSS property with the supported `overflow-wrap`
CSS property.