[OSDEV-1920] Update the Code Freeze section of the `RELEASE-PROTOCOL.md`. (#609)
[OSDEV-1920](https://opensupplyhub.atlassian.net/browse/OSDEV-1920) -
Releases. Update release protocol to support flexible code freeze and
improve Preprod stability checks.

In this PR, the **Code Freeze** section of the `RELEASE-PROTOCOL.md`
file was updated to include information about the flexible code freeze
process and the requirement to verify the stability of the Pre-prod
environment during deployment.

[OSDEV-1920]:
https://opensupplyhub.atlassian.net/browse/OSDEV-1920?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ