[OSDEV-1947] Introduced support for RBA deployment, disabled CSV export in RBA, and fixed FE environment check. (#614)
* Introduced infrastructure changes to support deployment of the RBA
environment. Included it in the CD pipelines in the same way as
Production and Staging, which means the RBA environment can now be
deployed via Git tags just like Production and Staging.
* Created the `export_csv_enabled` infrastructure switch to disable the
Amazon EventBridge scheduler for the CSV export job in the RBA
environment, as exporting the database to CSV is not needed in that
environment.
* Fixed a bug related to an incorrect environment check in the FE app,
which attempted to identify the local environment by comparing it with
`development`. However, the environment had been renamed to `local` some
time ago and was no longer passed as `development` from the BE during
local development