Removed unnecessary tag check and assignment of None in the "Deploy to AWS" GitHub workflow (#621)
Updated the `Deploy to AWS` GitHub workflow to rely on the required
`deploy-env` input instead of inferring the `environment` from the tag
name. Previously, this approach was necessary because the `[Release]
Deploy` workflow did not trigger `Deploy to AWS` via the API with the
specified `environment` - it only created a tag that triggered the
workflow automatically. With the latest changes, the `environment` is
now explicitly passed in the `[Release] Deploy` workflow. It has also
been confirmed that no branch or input combination can result in an
undefined `environment`. As a result, the fallback to `None` for the
`environment` value was removed from the `Deploy to AWS` workflow.

Tested by running `[Release] Deploy` workflow:

1.
https://github.com/opensupplyhub/open-supply-hub/actions/runs/14977617833
2.
https://github.com/opensupplyhub/open-supply-hub/actions/runs/14977623299