[OSDEV-1122] Update behavior of "Suggested Edit" to SLC entry point. (#628)
[OSDEV-1122](https://opensupplyhub.atlassian.net/browse/OSDEV-1122) -
SLC. Update behavior of "Suggested Edit" to SLC entry point.

The following changes were made:

- Updated the behavior of the `Suggest an Edit` button on production
location profile pages. The button now opens the Production Location
Information page of the SLC workflow in a new tab, allowing users to
suggest edits without having to re-search for the facility.
- Fixed redirection for unauthenticated users.If a user is not logged in
and clicks `Suggest an Edit` button they are now redirected to the
correct workflow step after logging in, instead of being taken to the
home page. This fix applies to all steps within the Contribute workflow.

[OSDEV-1122]:
https://opensupplyhub.atlassian.net/browse/OSDEV-1122?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ