[OSDEV-2032] Fixed the incorrect indexing of the location type in the production locations OpenSearch index (#669)
Fixed the incorrect indexing of the location type in the production
locations OpenSearch index. Previously, it was incorrectly taking the
processing type value as the location type — using the fourth item in
the `matched_values` array instead of the third, which contains the
location type. Also, updated the Logstash filters for both processing
type and location type to return only unique values and write them to
the production locations OpenSearch index.