[OSDEV-2266] [Partner Fields Enhancement] Add JSON Schema support for the Object type field (#816)
**[OSDEV-2266](https://opensupplyhub.atlassian.net/browse/OSDEV-2266)
[Partner Fields Enhancement] Add JSON Schema support for the Object type
field**

- Added JSON schema validation support for object type partner fields.
- Partner fields with object type can now have an associated JSON schema
that validates the structure and content of contributed data, providing
detailed error messages for validation failures.
- `0186_add_json_schema_to_partner_field.py` migration adds a
`json_schema` JSONField to the `PartnerField` model, allowing
administrators to define JSON schemas for validating object type partner
fields. The field is optional and is used to validate the structure and
content of contributed data when the partner field type is `object`.

[OSDEV-2266]:
https://opensupplyhub.atlassian.net/browse/OSDEV-2266?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ