[OSDEV-2047] Deleted Terraform configs and references for deprecated ContriCleaner service (#837)
* Removed all Terraform configurations and ECS service definitions
related to the deprecated standalone ContriCleaner service. Cleaned up
the repository by deleting unused code and references, as ContriCleaner
now operates exclusively as an internal Django library.
* Previously, there were two security groups with the same tags: one for
the Django app and another for ContriCleaner. After removing the
ContriCleaner service infrastructure, a bug was eliminated in which the
Django CLI task in the Development environment selected the wrong
security group - the one without database access, belonging to
ContriCleaner - which prevented Django management commands from running
against the database in the Development environment.